Error when max int or min int

+++ push_swap/libft/ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: valmpani <valmpani@student.42wolfsburg.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/12 10:59:56 by valmpani          #+#    #+#             */
-/*   Updated: 2023/05/12 11:05:54 by valmpani         ###   ########.fr       */
+/*   Updated: 2023/06/07 10:50:17 by valmpani         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long int	ft_atoi(const char *str)
 {
-	int	index;
-	int	minus;
-	int	number;
+	int			index;
+	int			minus;
+	long int	number;
 
 	index = 0;
 	minus = 1;
